opsplitsing van httpStudent
