+++ frontend/src/App.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { RouterProvider } from "react-router-dom";
 import { QueryClientProvider } from "@tanstack/react-query";
-import { queryClient } from "./util/student/httpStudent";
 import { router } from "./routes/router";
+import { queryClient } from "./util/student/config";
 
 const App = () => (
   <QueryClientProvider client={queryClient}>
