+++ frontend/src/pages/student/SignupStudent.tsx
@@ -10,10 +10,10 @@ import {
 } from '../../util/shared/validation';
 import Container from '../../components/shared/Container';
 import BoxBorder from '../../components/shared/BoxBorder';
-import { signupStudent } from '../../util/student/httpStudent';
 import LoadingIndicatorButton from '../../components/shared/LoadingIndicatorButton';
 import PrimaryButton from '../../components/shared/PrimaryButton';
 import { useTranslation } from 'react-i18next';
+import { signupStudent } from '@/util/student/auth';
 
 interface SignupFormData {
   firstName: string;
